more test cases, fixes and few more fixes

+++ src/main/java/com/phonepe/growth/mustang/composition/Composition.java
@@ -28,7 +28,7 @@ public abstract class Composition {
 
     public abstract boolean evaluate(EvaluationContext context);
 
-    public abstract long getScore(EvaluationContext context);
+    public abstract double getScore(EvaluationContext context);
 
     public abstract <T> T accept(CompositionVisitor<T> visitor);
 }
